clickhouse: rename async_object_storage to object_storage
There isn't anything async in this module.

[DDB-951]
